Add face expression detection
Add face-api.JS.

Add models for face API JS.

Add sample images for testing and developing face API.

Next, get around the issue "[plugin:vite:import-analysis] Failed to parse source for import analysis..." to import and load the face model. The models do not have file extension but the vite is expecting that it must be tsx and some sort. A resolve to be worked on next.